（1）加入了权限模块的Annotation （2）修改了API中时间格式的定义 （3）根据阿里规约规范代码格式

+++ coupon/src/main/java/cn/edu/xmu/oomall/coupon/controller/CouponController.java
@@ -8,6 +8,9 @@ import cn.edu.xmu.oomall.core.util.ReturnObject;
 import cn.edu.xmu.oomall.coupon.model.bo.CouponActivity;
 import cn.edu.xmu.oomall.coupon.model.vo.CouponActivityVo;
 import cn.edu.xmu.oomall.coupon.service.CouponService;
+import cn.edu.xmu.privilegegateway.annotation.annotation.Audit;
+import cn.edu.xmu.privilegegateway.annotation.annotation.LoginName;
+import cn.edu.xmu.privilegegateway.annotation.annotation.LoginUser;
 import com.github.pagehelper.PageInfo;
 import io.swagger.annotations.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -43,16 +46,21 @@ public class CouponController {
             @ApiResponse(code = 500, message = "服务器内部错误"),
             @ApiResponse(code = 504, message = "资源不存在")
     })
-    @RequestMapping(value="/couponactivities/{id}/products", method=RequestMethod.GET)
+    @RequestMapping(value = "/couponactivities/{id}/products", method = RequestMethod.GET)
     public Object listProductsByCouponActivityId(@ApiParam(value = "优惠活动ID", required = true) @PathVariable("id") Long couponActivityId,
-                                                @ApiParam(value = "页码") @RequestParam(value = "page", required = false, defaultValue = "1") Integer pageNumber,
-                                                @ApiParam(value = "每页数目") @RequestParam(value = "pageSize", required = false, defaultValue = "10") Integer pageSize) {
+                                                 @ApiParam(value = "页码") @RequestParam(value = "page", required = false, defaultValue = "1") Integer pageNumber,
+                                                 @ApiParam(value = "每页数目") @RequestParam(value = "pageSize", required = false, defaultValue = "10") Integer pageSize) {
+        if (couponActivityId < 0) {
+            return Common.decorateReturnObject(new ReturnObject<>(ReturnNo.FIELD_NOTVALID, ReturnNo.FIELD_NOTVALID.getMessage()));
+        }
+        pageNumber = pageNumber >= 1 ? pageNumber : 1;
+        pageSize = pageSize >= 1 ? pageSize : 10;
+        
         ReturnObject<PageInfo<VoObject>> retVoObject =
                 couponService.listProductsByCouponActivityId(couponActivityId, pageNumber, pageSize);
         return Common.decorateReturnObject(Common.getPageRetObject(retVoObject));
 
     }
-
 
 
     @ApiOperation(value = "查看商品的上线的优惠活动")
@@ -66,17 +74,24 @@ public class CouponController {
             @ApiResponse(code = 500, message = "服务器内部错误"),
             @ApiResponse(code = 504, message = "资源不存在")
     })
-    @RequestMapping(value = "products/{id}/couponactivities", method=RequestMethod.GET)
+    @RequestMapping(value = "products/{id}/couponactivities", method = RequestMethod.GET)
     public Object listCouponActivitiesByProductId(@ApiParam(value = "货品ID", required = true) @PathVariable("id") Long productId,
                                                   @ApiParam(value = "页码") @RequestParam(value = "page", required = false, defaultValue = "1") Integer pageNumber,
                                                   @ApiParam(value = "每页数目") @RequestParam(value = "pageSize", required = false, defaultValue = "10") Integer pageSize) {
+        if (productId < 0) {
+            return Common.decorateReturnObject(new ReturnObject<>(ReturnNo.FIELD_NOTVALID, ReturnNo.FIELD_NOTVALID.getMessage()));
+        }
+        pageNumber = pageNumber >= 1 ? pageNumber : 1;
+        pageSize = pageSize >= 1 ? pageSize : 10;
+
         ReturnObject<PageInfo<VoObject>> retVoObject =
                 couponService.listCouponActivitiesByProductId(productId, pageNumber, pageSize);
 
-        if (retVoObject.getCode().equals(ReturnNo.OK))
+        if (retVoObject.getCode().equals(ReturnNo.OK)) {
             return Common.decorateReturnObject(Common.getPageRetObject(retVoObject));
-        else
+        } else {
             return Common.decorateReturnObject(retVoObject);
+        }
     }
 
 
@@ -93,23 +108,28 @@ public class CouponController {
             @ApiResponse(code = 947, message = "开始时间不能晚于结束时间"),
             @ApiResponse(code = 950, message = "优惠卷领卷时间晚于活动开始时间")
     })
-    @RequestMapping(value = "/shops/{shopId}/couponactivities/{id}", method=RequestMethod.PUT)
+    @RequestMapping(value = "/shops/{shopId}/couponactivities/{id}", method = RequestMethod.PUT)
+    @Audit(departName = "shops")
     public Object updateCouponActivity(@ApiParam(value = "商店ID", required = true) @PathVariable("shopId") Long shopId,
                                        @ApiParam(value = "优惠活动ID", required = true) @PathVariable("id") Long couponActivityId,
                                        @ApiParam(value = "可修改的优惠活动信息", required = true) @Validated @RequestBody CouponActivityVo couponActivityVo,
-                                       Long userId, String userName, BindingResult bindingResult) {
-        userId = 123L;
-        userName = "hqg";
+                                       @LoginUser Long userId, @LoginName String userName, BindingResult bindingResult) {
+        if (shopId < 0 || couponActivityId < 0) {
+            return Common.decorateReturnObject(new ReturnObject<>(ReturnNo.FIELD_NOTVALID, ReturnNo.FIELD_NOTVALID.getMessage()));
+        }
+
         Object object = Common.processFieldErrors(bindingResult, httpServletResponse);
         if (object != null) {
             return object;
         }
 
-        if (couponActivityVo.getBeginTime().isAfter(couponActivityVo.getEndTime()))
+        if (couponActivityVo.getBeginTime().isAfter(couponActivityVo.getEndTime())) {
             return Common.decorateReturnObject(new ReturnObject<>(ReturnNo.ACT_LATE_BEGINTIME));
+        }
 
-        if (couponActivityVo.getCouponTime().isAfter((couponActivityVo.getBeginTime())))
+        if (couponActivityVo.getCouponTime().isAfter((couponActivityVo.getBeginTime()))) {
             return Common.decorateReturnObject(new ReturnObject<>(ReturnNo.COUPON_LATE_COUPONTIME));
+        }
 
         ReturnObject returnObject = couponService.updateCouponActivity(userId, userName, shopId, couponActivityId, couponActivityVo, null);
         return Common.decorateReturnObject(returnObject);
@@ -126,12 +146,15 @@ public class CouponController {
             @ApiResponse(code = 200, message = "成功"),
             @ApiResponse(code = 507, message = "当前状态禁止此操作"),
     })
-    @RequestMapping(value = "/shops/{shopId}/couponactivities/{id}", method=RequestMethod.DELETE)
+    @RequestMapping(value = "/shops/{shopId}/couponactivities/{id}", method = RequestMethod.DELETE)
+    @Audit(departName = "shops")
     public Object deleteCouponActivity(@ApiParam(value = "商店ID", required = true) @PathVariable("shopId") Long shopId,
                                        @ApiParam(value = "优惠活动ID", required = true) @PathVariable("id") Long couponActivityId,
-                                       Long userId, String userName) {
-        userId = 123L;
-        userName = "hqg";
+                                       @LoginUser Long userId, @LoginName String userName) {
+        if (shopId < 0 || couponActivityId < 0) {
+            return Common.decorateReturnObject(new ReturnObject<>(ReturnNo.FIELD_NOTVALID, ReturnNo.FIELD_NOTVALID.getMessage()));
+        }
+
         ReturnObject returnObject = couponService.deleteCouponActivity(userId, userName, shopId, couponActivityId);
         return Common.decorateReturnObject(returnObject);
     }
@@ -147,13 +170,16 @@ public class CouponController {
             @ApiResponse(code = 200, message = "成功"),
             @ApiResponse(code = 507, message = "当前状态禁止此操作"),
     })
-    @RequestMapping(value = "/shops/{shopId}/couponactivities/{id}/onsales/{sid}", method=RequestMethod.POST)
+    @RequestMapping(value = "/shops/{shopId}/couponactivities/{id}/onsales/{sid}", method = RequestMethod.POST)
+    @Audit(departName = "shops")
     public Object insertCouponOnsale(@ApiParam(value = "商店ID", required = true) @PathVariable("shopId") Long shopId,
-                                                         @ApiParam(value = "优惠活动ID", required = true) @PathVariable("id") Long couponActivityId,
-                                                         @ApiParam(value = "销售活动ID", required = true) @PathVariable("sid") Long onsaleId,
-                                                         Long userId, String userName){
-        userId = 123L;
-        userName = "hqg";
+                                     @ApiParam(value = "优惠活动ID", required = true) @PathVariable("id") Long couponActivityId,
+                                     @ApiParam(value = "销售活动ID", required = true) @PathVariable("sid") Long onsaleId,
+                                     @LoginUser Long userId, @LoginName String userName) {
+        if (shopId < 0 || couponActivityId < 0 || onsaleId < 0) {
+            return Common.decorateReturnObject(new ReturnObject<>(ReturnNo.FIELD_NOTVALID, ReturnNo.FIELD_NOTVALID.getMessage()));
+        }
+
         ReturnObject returnObject = couponService.insertCouponOnsale(userId, userName, shopId, couponActivityId, onsaleId);
         return Common.decorateReturnObject(returnObject);
     }
@@ -169,12 +195,15 @@ public class CouponController {
             @ApiResponse(code = 200, message = "成功"),
             @ApiResponse(code = 507, message = "当前状态禁止此操作"),
     })
-    @RequestMapping(value = "/shops/{shopId}/coupononsale/{id}", method=RequestMethod.DELETE)
+    @RequestMapping(value = "/shops/{shopId}/coupononsale/{id}", method = RequestMethod.DELETE)
+    @Audit(departName = "shops")
     public Object deleteCouponOnsale(@ApiParam(value = "商店ID", required = true) @PathVariable("shopId") Long shopId,
                                      @ApiParam(value = "couponOnsaleId", required = true) @PathVariable("id") Long couponOnsaleId,
-                                     Long userId, String userName){
-        userId = 123L;
-        userName = "hqg";
+                                     @LoginUser Long userId, @LoginName String userName) {
+        if (shopId < 0 || couponOnsaleId < 0) {
+            return Common.decorateReturnObject(new ReturnObject<>(ReturnNo.FIELD_NOTVALID, ReturnNo.FIELD_NOTVALID.getMessage()));
+        }
+
         ReturnObject returnObject = couponService.deleteCouponOnsale(userId, userName, shopId, couponOnsaleId);
         return Common.decorateReturnObject(returnObject);
     }
@@ -190,12 +219,15 @@ public class CouponController {
             @ApiResponse(code = 200, message = "成功"),
             @ApiResponse(code = 507, message = "当前状态禁止此操作"),
     })
-    @RequestMapping(value = "/shops/{shopId}/couponactivities/{id}/online", method=RequestMethod.PUT)
+    @RequestMapping(value = "/shops/{shopId}/couponactivities/{id}/online", method = RequestMethod.PUT)
+    @Audit(departName = "shops")
     public Object updateCouponActivityToOnline(@ApiParam(value = "商店ID", required = true) @PathVariable("shopId") Long shopId,
                                                @ApiParam(value = "优惠活动ID", required = true) @PathVariable("id") Long couponActivityId,
-                                               Long userId, String userName) {
-        userId = 123L;
-        userName = "hqg";
+                                               @LoginUser Long userId, @LoginName String userName) {
+        if (shopId < 0 || couponActivityId < 0) {
+            return Common.decorateReturnObject(new ReturnObject<>(ReturnNo.FIELD_NOTVALID, ReturnNo.FIELD_NOTVALID.getMessage()));
+        }
+
         ReturnObject returnObject = couponService.updateCouponActivity(userId, userName, shopId, couponActivityId, null, CouponActivity.State.ONLINE);
         return Common.decorateReturnObject(returnObject);
     }
@@ -211,15 +243,16 @@ public class CouponController {
             @ApiResponse(code = 200, message = "成功"),
             @ApiResponse(code = 507, message = "当前状态禁止此操作"),
     })
-    @RequestMapping(value = "/shops/{shopId}/couponactivities/{id}/offline", method=RequestMethod.PUT)
+    @RequestMapping(value = "/shops/{shopId}/couponactivities/{id}/offline", method = RequestMethod.PUT)
+    @Audit(departName = "shops")
     public Object updateCouponActivityToOffline(@ApiParam(value = "商店ID", required = true) @PathVariable("shopId") Long shopId,
-                                               @ApiParam(value = "优惠活动ID", required = true) @PathVariable("id") Long couponActivityId,
-                                               Long userId, String userName) {
-        userId = 123L;
-        userName = "hqg";
+                                                @ApiParam(value = "优惠活动ID", required = true) @PathVariable("id") Long couponActivityId,
+                                                @LoginUser Long userId, @LoginName String userName) {
+        if (shopId < 0 || couponActivityId < 0) {
+            return Common.decorateReturnObject(new ReturnObject<>(ReturnNo.FIELD_NOTVALID, ReturnNo.FIELD_NOTVALID.getMessage()));
+        }
+
         ReturnObject returnObject = couponService.updateCouponActivity(userId, userName, shopId, couponActivityId, null, CouponActivity.State.OFFLINE);
         return Common.decorateReturnObject(returnObject);
     }
-
-
 }
